Extract common helper functions into its own file.
